backend/requirements.txt
anthropic
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
ruff